+++ changelog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [Unreleased]
+
+### Added
+
+- Add `drawLinks`, `getAnchorsPosition` and `getDefaultAnchor` methods in `DefaultDrawer.js`.
+- Create a new `draw` method that executes the `drawLinks` and `drawComponents` methods.
+
+### Changed
+
+- Rename the `draw` method to `drawComponents`.
+
 ## [0.8.1] - 2022-10-12
 
 ### Fixed